Update group of animals.cpp
Explanation in Simple Terms:

1. Ask for the Number of Animals:

The program first asks the user how many domestic animals (N) and wild animals (X) they want to enter.

2. Store Animal Names:

Two arrays (domesticAnimals and wildAnimals) are created to store the names of the animals. The size of each array depends on the values of N and X.

3. Input Animal Names:

The program uses a loop to ask the user to enter the names of the domestic animals one by one.

Then, it uses another loop to ask for the names of the wild animals.

4. Print All Animals:

After collecting all the names, the program prints all the animals. It first prints the domestic animals and then the wild animals.