Add files via upload
this is our SACCO code which apparently is advanced compared to what we know about coding , this is all we could come up with. there are alaot of ways to improve this code like adding a User interface , database and more other futures. Our code accepts registration signup and also signin and then allow the user to go to the main menu of depositing , withdraw, view statements and also changing password